requirements.txt
streamlit
faiss-cpu
sentence-transformers
ctransformers
pypdf
python-docx
langchain
torch
pydantic